Projects/VAE/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.1.0
numpy>=1.21.0